refactor(scoring): split the god module into response builders and rescoring
Phase 2 of the ScoringReader decomposition, mechanical now that #989
retired the namespace-patch coupling.

- _deps.py: ScoringDeps bundle (shared by the submodules without
  circular imports).
- _response_builders.py: the SQL-grade-table and eval-JSON-file payload
  builders (_severity_bucket .. _build_response_from_eval_files), moved
  verbatim. The WS6 serialization ratchet's wire-boundary declaration
  moves with the to_camel_dict call sites.
- _rescoring.py: the accumulated-rescore machinery with the partial-
  coverage persist gate (_rescore_runs_by_dimension ..
  _rescore_accumulated_response), moved verbatim.
- __init__.py shrinks from 686 to ~380 lines: the public API
  (get_scores_raw/slim, scored_run_dimensions, rescore_accumulated,
  get_project_scores), the fetcher factories, and facade re-exports of
  every moved name, so external callers and the one remaining
  orchestration-glue patch target are unchanged.

No behavior change; the existing suite is the net.
