+++ src/quodeq/services/scoring/_deps.py
@@ -1,0 +1,30 @@
+"""Injectable dependency bundle for the scoring reader (see ScoringDeps)."""
+from __future__ import annotations
+
+from dataclasses import dataclass
+from typing import Callable
+
+
+@dataclass(frozen=True)
+class ScoringDeps:
+    """Injectable dependency bundle for the scoring reader.
+
+    A ``None`` field resolves to the production callable at call time, so
+    the seam is purely additive: existing callers pass nothing and see no
+    change. Tests construct a ``ScoringDeps`` with fakes instead of
+    patching this module's attributes — the namespace-patch coupling is
+    what made the previous decomposition attempt revert.
+    """
+
+    read_run_data: Callable | None = None
+    read_run_scalars: Callable | None = None
+    dismissed_keys: Callable | None = None
+    deleted_keys: Callable | None = None
+    cached_accumulated: Callable | None = None
+    rescore_dimension: Callable | None = None
+    rescore_runs_by_dimension: Callable | None = None
+    recompute_summary: Callable | None = None
+
+
+_NO_DEPS = ScoringDeps()
+
